+++ src/main/java/entities/Defender.java
@@ -1,8 +1,10 @@
+package entities;
+
 import java.util.HashMap;
 
-public class Goalkeeper extends Player {
-    public Goalkeeper(String name, int age, double height, double weight, String team, int rating,
-                      double value, String position, HashMap<String, Integer> skills) {
+public class Defender extends Player {
+    public Defender(String name, int age, double height, double weight, String team, int rating,
+                    double value, String position, HashMap<String, Integer> skills) {
         super(name, age, height, weight, team, rating, value, position, skills);
     }
 
@@ -11,3 +13,5 @@ public class Goalkeeper extends Player {
         System.out.println("This player's position is " + super.getPosition());
     }
 }
+
+
