+++ src/main/java/entities/Forward.java
@@ -1,9 +1,12 @@
+package entities;
+
 import java.util.HashMap;
 
-public class Midfielder extends Player {
-    public Midfielder(String name, int age, double height, double weight, String team, int rating,
-                      double value, String position, HashMap<String, Integer> skills) {
+public class Forward extends Player {
+    public Forward(String name, int age, double height, double weight, String team, int rating,
+                   double value, String position, HashMap<String, Integer> skills) {
         super(name, age, height, weight, team, rating, value, position, skills);
+
 
     }
