+++ src/main/java/entities/Midfielder.java
@@ -1,9 +1,12 @@
+package entities;
+
 import java.util.HashMap;
 
-public class Defender extends Player {
-    public Defender(String name, int age, double height, double weight, String team, int rating,
-                    double value, String position, HashMap<String, Integer> skills) {
+public class Midfielder extends Player {
+    public Midfielder(String name, int age, double height, double weight, String team, int rating,
+                      double value, String position, HashMap<String, Integer> skills) {
         super(name, age, height, weight, team, rating, value, position, skills);
+
     }
 
     @Override
@@ -11,5 +14,3 @@ public class Defender extends Player {
         System.out.println("This player's position is " + super.getPosition());
     }
 }
-
-
