+++ src/main/java/entities/Player.java
@@ -1,3 +1,5 @@
+package entities;
+
 import java.util.HashMap;
 
 /*
