+++ src/main/java/entities/PlayerFactory.java
@@ -1,3 +1,5 @@
+package entities;
+
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Set;
