+++ src/main/java/io/InputData.java
@@ -1,3 +1,5 @@
+package io;
+
 import java.io.IOException;
 
 public interface InputData {
