+++ src/main/java/io/InputPlayerName.java
@@ -1,3 +1,7 @@
+package io;
+
+import search.SearchForPlayer;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
