+++ src/main/java/io/PlayerPresenter.java
@@ -1,3 +1,5 @@
+package io;
+
 class PlayerPresenter {
 
 }
