+++ src/main/java/io/PlayerPropertiesIterator.java
@@ -1,3 +1,5 @@
+package io;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.*;
