+++ src/main/java/search/SearchByPlayerAttributes.java
@@ -1,3 +1,8 @@
+package search;
+
+import data.PlayerDatabase;
+import entities.Player;
+
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
