+++ src/main/java/search/SearchForPlayer.java
@@ -1,3 +1,7 @@
+package search;
+
+import data.PlayerDatabase;
+import entities.Player;
 import org.apache.commons.text.similarity.LevenshteinDistance;
 
 import java.util.ArrayList;
