+++ src/main/java/services/CSVAdapter.java
@@ -1,7 +1,14 @@
+package services;
+
 import com.opencsv.CSVReader;
 import com.opencsv.CSVReaderBuilder;
+import data.PlayerDatabase;
+import data.TeamDatabase;
+import entities.Player;
+import entities.PlayerFactory;
 
 import java.io.FileReader;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -111,7 +118,7 @@ public class CSVAdapter implements InputAdapter {
      * @param databaseFile the path of the database file (CSV)
      */
     @Override
-    public void dataDump(String databaseFile) {
+    public void dataDump(String databaseFile) throws IOException {
         try {
             // Create fileReader and CsvReader objects
             FileReader fileReader = new FileReader(databaseFile);
