+++ src/test/java/PlayerStatsCalculatorTest.java
@@ -1,3 +1,5 @@
+import data.PlayerStatsCalculator;
+import entities.Player;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
