+++ src/test/java/PlayersPresenterTest.java
@@ -1,3 +1,5 @@
+import entities.Player;
+import io.PlayersPresenter;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
