+++ src/test/java/TestTeam.java
@@ -1,3 +1,5 @@
+import entities.Player;
+import entities.Team;
 import org.junit.Before;
 import org.junit.Test;
