+++ src/test/java/TestTeamDatabase.java
@@ -1,3 +1,6 @@
+import data.TeamDatabase;
+import entities.Player;
+import entities.Team;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
